+++ jkube-kit/enricher/api/src/test/java/org/eclipse/jkube/kit/enricher/handler/PodTemplateHandlerTest.java
@@ -20,38 +20,37 @@ import org.eclipse.jkube.kit.common.JavaProject;
 import org.eclipse.jkube.kit.config.resource.GroupArtifactVersion;
 import org.eclipse.jkube.kit.config.resource.ResourceConfig;
 import org.eclipse.jkube.kit.config.resource.VolumeConfig;
-import mockit.Mocked;
 import org.junit.Before;
 import org.junit.Test;
 
 import java.util.ArrayList;
 import java.util.List;
 
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
+import static org.mockito.Mockito.RETURNS_DEEP_STUBS;
+import static org.mockito.Mockito.mock;
 
 public class PodTemplateHandlerTest {
 
-    @Mocked
-    ProbeHandler probeHandler;
-
-    @Mocked
-    JavaProject project;
-
-    List<String> mounts = new ArrayList<>();
-    List<VolumeConfig> volumes1 = new ArrayList<>();
-
-    List<ImageConfiguration> images = new ArrayList<>();
-
-    List<String> ports = new ArrayList<>();
-
-    List<String> tags = new ArrayList<>();
+    private ProbeHandler probeHandler;
+    private JavaProject project;
+    private PodTemplateHandler podTemplateHandler;
+    private List<String> mounts;
+    private List<VolumeConfig> volumes1;
+    private List<ImageConfiguration> images;
 
     @Before
-    public void before(){
+    public void before() {
+        mounts = new ArrayList<>();
+        volumes1 = new ArrayList<>();
+        images = new ArrayList<>();
+        List<String> ports = new ArrayList<>();
+        List<String> tags = new ArrayList<>();
 
         //volume config with name and multiple mount
         mounts.add("/path/system");
@@ -73,15 +72,14 @@ public class PodTemplateHandlerTest {
                 .registry("docker.io").build();
 
         images.add(imageConfiguration);
+        probeHandler = mock(ProbeHandler.class, RETURNS_DEEP_STUBS);
+        project = mock(JavaProject.class, RETURNS_DEEP_STUBS);
+        ContainerHandler containerHandler = getContainerHandler();
+        podTemplateHandler = new PodTemplateHandler(containerHandler);
     }
 
     @Test
     public void podWithoutVolumeTemplateHandlerTest() {
-
-        ContainerHandler containerHandler = getContainerHandler();
-
-        PodTemplateHandler podTemplateHandler = new PodTemplateHandler(containerHandler);
-
         //Pod without Volume Config
         ResourceConfig config = ResourceConfig.builder()
                 .imagePullPolicy("IfNotPresent")
@@ -90,7 +88,7 @@ public class PodTemplateHandlerTest {
                 .replicas(5)
                 .build();
 
-        PodTemplateSpec podTemplateSpec = podTemplateHandler.getPodTemplate(config, images);
+        PodTemplateSpec podTemplateSpec = podTemplateHandler.getPodTemplate(config, null, images);
 
         //Assertion
         assertEquals("test-account", podTemplateSpec.getSpec().getServiceAccountName());
@@ -104,15 +102,8 @@ public class PodTemplateHandlerTest {
                 .getContainers().get(0).getImagePullPolicy());
     }
 
-    private ContainerHandler getContainerHandler() {
-        return new ContainerHandler(project.getProperties(), new GroupArtifactVersion("g","a","v"), probeHandler);
-    }
-
     @Test
     public void podWithEmotyVolumeTemplateHandlerTest(){
-        ContainerHandler containerHandler = getContainerHandler();
-
-        PodTemplateHandler podTemplateHandler = new PodTemplateHandler(containerHandler);
         //Pod with empty Volume Config and wihtout ServiceAccount
         ResourceConfig config = ResourceConfig.builder()
                 .imagePullPolicy("IfNotPresent")
@@ -121,7 +112,7 @@ public class PodTemplateHandlerTest {
                 .volumes(volumes1)
                 .build();
 
-        PodTemplateSpec podTemplateSpec = podTemplateHandler.getPodTemplate(config,images);
+        PodTemplateSpec podTemplateSpec = podTemplateHandler.getPodTemplate(config, null, images);
 
         //Assertion
         assertNull(podTemplateSpec.getSpec().getServiceAccountName());
@@ -131,9 +122,6 @@ public class PodTemplateHandlerTest {
 
     @Test
     public void podWithVolumeTemplateHandlerTest(){
-        ContainerHandler containerHandler = getContainerHandler();
-
-        PodTemplateHandler podTemplateHandler = new PodTemplateHandler(containerHandler);
         //Config with Volume Config and ServiceAccount
         //valid type
         VolumeConfig volumeConfig1 = VolumeConfig.builder().name("test")
@@ -149,7 +137,7 @@ public class PodTemplateHandlerTest {
                 .volumes(volumes1)
                 .build();
 
-        PodTemplateSpec podTemplateSpec = podTemplateHandler.getPodTemplate(config,images);
+        PodTemplateSpec podTemplateSpec = podTemplateHandler.getPodTemplate(config, null, images);
 
         //Assertion
         assertEquals("test-account",podTemplateSpec.getSpec().getServiceAccountName());
@@ -163,10 +151,6 @@ public class PodTemplateHandlerTest {
 
     @Test
     public void podWithInvalidVolumeTypeTemplateHandlerTest(){
-        ContainerHandler containerHandler = getContainerHandler();
-
-        PodTemplateHandler podTemplateHandler = new PodTemplateHandler(containerHandler);
-
         //invalid type
         VolumeConfig volumeConfig1 = VolumeConfig.builder().name("test")
                 .mounts(mounts).type("hoStPath").path("/test/path").build();
@@ -181,7 +165,7 @@ public class PodTemplateHandlerTest {
                 .volumes(volumes1)
                 .build();
 
-        PodTemplateSpec podTemplateSpec = podTemplateHandler.getPodTemplate(config,images);
+        PodTemplateSpec podTemplateSpec = podTemplateHandler.getPodTemplate(config, null, images);
 
         //Assertion
         assertEquals("test-account",podTemplateSpec.getSpec().getServiceAccountName());
@@ -191,10 +175,6 @@ public class PodTemplateHandlerTest {
 
     @Test
     public void podWithoutEmptyTypeTemplateHandlerTest(){
-        ContainerHandler containerHandler = getContainerHandler();
-
-        PodTemplateHandler podTemplateHandler = new PodTemplateHandler(containerHandler);
-
         //empty type
         VolumeConfig volumeConfig1 = VolumeConfig.builder().name("test").mounts(mounts).build();
         volumes1.clear();
@@ -208,11 +188,28 @@ public class PodTemplateHandlerTest {
                 .volumes(volumes1)
                 .build();
 
-        PodTemplateSpec podTemplateSpec = podTemplateHandler.getPodTemplate(config,images);
+        PodTemplateSpec podTemplateSpec = podTemplateHandler.getPodTemplate(config, null, images);
 
         //Assertion
         assertEquals("test-account",podTemplateSpec.getSpec().getServiceAccountName());
         assertTrue(podTemplateSpec.getSpec().getVolumes().isEmpty());
         assertNotNull(podTemplateSpec.getSpec().getContainers());
+    }
+
+    @Test
+    public void getPodTemplate_withRestartPolicyAndResourceConfig_shouldGeneratePodTemplateWithConfiguredRestartPolicy() {
+        // Given
+        ResourceConfig config = ResourceConfig.builder().build();
+
+        // When
+        PodTemplateSpec podTemplateSpec = podTemplateHandler.getPodTemplate(config, "Always", images);
+
+        // Then
+        assertThat(podTemplateSpec)
+            .hasFieldOrPropertyWithValue("spec.restartPolicy", "Always");
+    }
+
+    private ContainerHandler getContainerHandler() {
+        return new ContainerHandler(project.getProperties(), new GroupArtifactVersion("g","a","v"), probeHandler);
     }
 }